Move native-related elements to the 'native' element group folder
