+++ angular/components/elements/element-group-native/ss-anchor/ss-anchor.component.scss
@@ -1,0 +1,4 @@
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
