+++ angular/components/elements/element-group-native/ss-button/models/ButtonTriggerType.ts
