+++ angular/components/elements/element-group-native/ss-button/ss-button.component.scss
@@ -1,7 +1,7 @@
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
 
 ss-button {
     @include flex(center, center, row);
