+++ angular/components/elements/element-group-native/ss-footer/ss-footer.component.ts
