+++ angular/components/elements/element-group-native/ss-input/ss-input.component.scss
@@ -1,8 +1,8 @@
 // ss-input.component.scss
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
 
 ss-input {
     @include flex(center, center, row);
