+++ angular/components/elements/element-group-native/ss-paragraph/ss-paragraph.component.scss
@@ -1,8 +1,8 @@
 // ss-paragraph.component.scss
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
 
 ss-paragraph {
     @include flex(center, center, column);
