+++ angular/components/elements/element-group-native/ss-select/ss-select.component.scss
@@ -1,8 +1,8 @@
 // ss-select.component.scss
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
 
 ss-select {
     @include flex(center, center, row);
