+++ angular/components/elements/element-group-native/ss-title/ss-title.component.ts
@@ -1,6 +1,6 @@
 // ss-title.component.ts 
 import {Component, Input} from "@angular/core";
-import {TagType} from "../../../models/html/TagType";
+import {TagType} from "../../../../models/html/TagType";
 
 @Component({
     selector: 'ss-title',
