--- angular/components/elements/ss-anchor/ss-anchor.component.scss
@@ -1,4 +1,0 @@
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
